Split factories into their own files
