+++ src/EditorFeatures/Core/ExternalAccess/VSTypeScript/VSTypeScriptLspServiceProvider.cs
@@ -10,7 +10,7 @@ using Microsoft.CodeAnalysis.LanguageServer;
 
 namespace Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript;
 
-[Export, Shared]
+[Export(typeof(VSTypeScriptLspServiceProvider)), Shared]
 internal class VSTypeScriptLspServiceProvider : AbstractLspServiceProvider
 {
     [ImportingConstructor]
