+++ src/EditorFeatures/Core/LanguageServer/Handlers/Completion/CompletionListCacheFactory.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Composition;
+using Microsoft.CodeAnalysis.Host.Mef;
+
+namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Completion;
+
+[ExportRoslynLspServiceFactory(typeof(CompletionListCache)), Shared]
+internal class CompletionListCacheFactory : ILspServiceFactory
+{
+    [ImportingConstructor]
+    [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+    public CompletionListCacheFactory()
+    {
+    }
+
+    public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind) => new CompletionListCache();
+}
