+++ src/EditorFeatures/Core/LanguageServer/Handlers/Completion/CompletionResolveHandlerFactory.cs
@@ -1,0 +1,31 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Composition;
+using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.LanguageServer.Handler.Completion;
+using Microsoft.CodeAnalysis.Options;
+
+namespace Microsoft.CodeAnalysis.LanguageServer.Handler
+{
+    [ExportRoslynLspServiceFactory(typeof(CompletionResolveHandler)), Shared]
+    internal sealed class CompletionResolveHandlerFactory : ILspServiceFactory
+    {
+        private readonly IGlobalOptionService _globalOptions;
+
+        [ImportingConstructor]
+        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+        public CompletionResolveHandlerFactory(IGlobalOptionService globalOptions)
+        {
+            _globalOptions = globalOptions;
+        }
+
+        public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
+        {
+            var completionListCache = lspServices.GetRequiredService<CompletionListCache>();
+            return new CompletionResolveHandler(_globalOptions, completionListCache);
+        }
+    }
+}
