+++ src/Features/LanguageServer/Protocol/Handler/Diagnostics/Experimental/ExperimentalDocumentPullDiagnosticsHandler.cs
@@ -24,32 +24,6 @@ using DocumentDiagnosticReport = SumType<FullDocumentDiagnosticReport, Unchanged
 // See https://github.com/microsoft/vscode-languageserver-node/blob/main/protocol/src/common/proposed.diagnostics.md#textDocument_diagnostic
 using DocumentDiagnosticPartialReport = SumType<SumType<FullDocumentDiagnosticReport, UnchangedDocumentDiagnosticReport>, DocumentDiagnosticPartialResult>;
 
-[ExportRoslynLspServiceFactory(typeof(ExperimentalDocumentPullDiagnosticsHandler)), Shared]
-internal class ExperimentalDocumentPullDiagnosticHandlerFactory : ILspServiceFactory
-{
-    private readonly IDiagnosticService _diagnosticService;
-    private readonly IDiagnosticAnalyzerService _analyzerService;
-    private readonly EditAndContinueDiagnosticUpdateSource _editAndContinueDiagnosticUpdateSource;
-    private readonly IGlobalOptionService _globalOptions;
-
-    [ImportingConstructor]
-    [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-    public ExperimentalDocumentPullDiagnosticHandlerFactory(
-        IDiagnosticService diagnosticService,
-        IDiagnosticAnalyzerService analyzerService,
-        EditAndContinueDiagnosticUpdateSource editAndContinueDiagnosticUpdateSource,
-        IGlobalOptionService globalOptions)
-    {
-        _diagnosticService = diagnosticService;
-        _analyzerService = analyzerService;
-        _editAndContinueDiagnosticUpdateSource = editAndContinueDiagnosticUpdateSource;
-        _globalOptions = globalOptions;
-    }
-
-    public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
-        => new ExperimentalDocumentPullDiagnosticsHandler(_diagnosticService, _analyzerService, _editAndContinueDiagnosticUpdateSource, _globalOptions);
-}
-
 [Method(ExperimentalMethods.TextDocumentDiagnostic)]
 internal class ExperimentalDocumentPullDiagnosticsHandler : AbstractPullDiagnosticHandler<DocumentDiagnosticParams, DocumentDiagnosticPartialReport, DocumentDiagnosticReport?>
 {
