+++ src/Features/LanguageServer/Protocol/Handler/Diagnostics/Experimental/ExperimentalWorkspacePullDiagnosticsHandler.cs
@@ -19,32 +19,6 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics.Experimental
 
 using WorkspaceDocumentDiagnosticReport = SumType<WorkspaceFullDocumentDiagnosticReport, WorkspaceUnchangedDocumentDiagnosticReport>;
 
-[ExportRoslynLspServiceFactory(typeof(ExperimentalWorkspacePullDiagnosticsHandler)), Shared]
-internal class ExperimentalWorkspacePullDiagnosticHandlerFactory : ILspServiceFactory
-{
-    private readonly IDiagnosticService _diagnosticService;
-    private readonly IDiagnosticAnalyzerService _analyzerService;
-    private readonly EditAndContinueDiagnosticUpdateSource _editAndContinueDiagnosticUpdateSource;
-    private readonly IGlobalOptionService _globalOptions;
-
-    [ImportingConstructor]
-    [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-    public ExperimentalWorkspacePullDiagnosticHandlerFactory(
-        IDiagnosticService diagnosticService,
-        IDiagnosticAnalyzerService analyzerService,
-        EditAndContinueDiagnosticUpdateSource editAndContinueDiagnosticUpdateSource,
-        IGlobalOptionService globalOptions)
-    {
-        _diagnosticService = diagnosticService;
-        _analyzerService = analyzerService;
-        _editAndContinueDiagnosticUpdateSource = editAndContinueDiagnosticUpdateSource;
-        _globalOptions = globalOptions;
-    }
-
-    public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
-        => new ExperimentalWorkspacePullDiagnosticsHandler(_diagnosticService, _analyzerService, _editAndContinueDiagnosticUpdateSource, _globalOptions);
-}
-
 [Method(ExperimentalMethods.WorkspaceDiagnostic)]
 internal class ExperimentalWorkspacePullDiagnosticsHandler : AbstractPullDiagnosticHandler<WorkspaceDiagnosticParams, WorkspaceDiagnosticReport, WorkspaceDiagnosticReport?>
 {
