+++ src/Features/LanguageServer/Protocol/Handler/RequestTelemetryLoggerFactory.cs
@@ -1,0 +1,24 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Composition;
+using Microsoft.CodeAnalysis.Host.Mef;
+
+namespace Microsoft.CodeAnalysis.LanguageServer.Handler;
+
+[ExportRoslynLspServiceFactory(typeof(RequestTelemetryLogger)), Shared]
+internal class RequestTelemetryLoggerFactory : ILspServiceFactory
+{
+    [ImportingConstructor]
+    [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+    public RequestTelemetryLoggerFactory()
+    {
+    }
+
+    public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
+    {
+        return new RequestTelemetryLogger(serverKind.ToTelemetryString());
+    }
+}
