+++ src/Features/LanguageServer/Protocol/IClientCapabilitiesProvider.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.VisualStudio.LanguageServer.Protocol;
+
+namespace Microsoft.CodeAnalysis.LanguageServer
+{
+    internal interface IClientCapabilitiesProvider : ILspService
+    {
+        ClientCapabilities GetClientCapabilities();
+    }
+}
